battle updates, fixed Necromancy skill
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@1650 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/battle2/battle_stats.h
@@ -221,7 +221,7 @@ namespace Battle2
 	u16	position;
 	u32	hp;
 	u32	count;
-	u16	dead;
+	u32	dead;
 	u8	shots;
 	u8	disruptingray;
 	bool	reflect;
